make a start on DB schema and Item query and insert
(some changes cherry-picked from https://github.com/guardian/pinboard/commit/69820c93d9645709afb919782366d8f651126d00)
